新增：激活武器特效

+++ Assets/Scripts/Game/Modules/Player/PlayerInventory.cs
@@ -237,7 +237,7 @@ public partial class PlayerManager
             );
     }
 
-    public long GetCurrWeaponLocation()
+    public int GetCurrWeaponLocation()
     {
         RefreshWeaponActive();
         return GetInventoryData().CurrentWeaponLocation;
